Approach2/requirements.txt
spacy
nltk
pandas[excel]
scikit-learn
biopython
fastapi[standard]
uvicorn